+++ modules/canvaskit/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `getShadowLocalBounds()` to estimate the bounds of the shadows drawn by `Canvas.drawShadow`.
  - compile.sh now takes "no_matrix", which will omit the helper JS to deal with 3x3, 4x4 and
    SkColorMatrix (in case clients have logic to deal with that themselves).
+ - `CanvasKit.RuntimeEffect.Make` now takes an optional callback function that will be called
+   with any compilation error.
 
 ### Breaking
  - `MakeImprovedNoise` is removed.